requirements-prod.txt
# ============================================
# Production Requirements (Vercel/Lambda compatible)
# ============================================

# Core Flask Framework
Flask
flask-cors
flask-jwt-extended

# Supabase
supabase

# Redis for Rate Limiting & Caching
redis

# Machine Learning & Data Processing
scikit-learn
numpy
pandas
opencv-python-headless
Pillow
joblib
tqdm

# Utilities
python-dotenv
PyJWT
passlib
bcrypt
click
python-jose

# Monitoring
sentry-sdk

# Production WSGI server
gunicorn
